==> 3_code/1.1_graph_simanfor_results.R <==
#------------------------------------------------------------------------------------------#
####                           Plot SIMANFOR biomass results                            ####
#                                                                                          #
#                            Aitor Vázquez Veloso, 31/03/2022                              #
#                              Last modification: 05/07/2024                               #
#------------------------------------------------------------------------------------------#


#### Basic steps ####

library(ggplot2)
library(viridis)
library(dplyr)
library(stringr)
library(ggpubr)

setwd('')


#### Load data ####

# load information
load('1_data/1_results/simulation_results.RData')

# rename to simplify
df <- stand_evolution

# delete growth extra period on control simulations
#df <- df[df$T <= 120, ]


#### Manage data before graph them ####

# # column for mixture
# df$Mixture <- ifelse(str_detect(df$n_scnr, 'Fsyl'), 'Psylvestris x Fsylvatica',
#                      ifelse(str_detect(df$n_scnr, 'Qpyr'), 'Psylvestris x Qpyrenaica',
#                                  ifelse(str_detect(df$n_scnr, 'Ppin'), 'Psylvestris x Ppinaster', 
#                                         'Psylvestris x Pnigra')))
# 
# # rename scenario labels
# df$n_scnr <- ifelse(str_detect(df$n_scnr, 'c12'), 'SI 12',
#                      ifelse(str_detect(df$n_scnr, 'c15'), 'SI 15',
#                             ifelse(str_detect(df$n_scnr, 'c18'), 'SI 18', 
#                                    ifelse(str_detect(df$n_scnr, 'c21'), 'SI 21',
#                                           ifelse(str_detect(df$n_scnr, 'c24'), 'SI 24', 'Control')))))


#### Graph Density among treatments ####

graph_N <- 
  ggplot(df, aes(x = T, y = N, 
                 group = n_scnr, colour = n_scnr)) +  # group by scnr
  
  # text
  labs(title = "Stand density evolution",  
       # subtitle = "",  
       x = "Stand age (years)",  
       y = "Density (trees/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        # plot.subtitle = element_text(size = 15, hjust = 0.5),  
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line()   # lines

# watch and save graph
graph_N
ggsave(filename = '4_figures/simulation_graphs/N.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph DG among treatments ####

graph_dg <- 
  ggplot(df, aes(x = T, y = dg, 
                 group = n_scnr, colour = n_scnr)) +  # group by scnr
  
  # text
  labs(title = "Stand quadratic mean diameter evolution",  
       # subtitle = "",  
       x = "Stand age (years)",  
       y = "Quadratic mean diameter (cm)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        # plot.subtitle = element_text(size = 15, hjust = 0.5),  
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line()  # lines

# watch and save graph
graph_dg
ggsave(filename = '4_figures/simulation_graphs/DG.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph G among treatments ####

graph_G <- 
  ggplot(df, aes(x = T, y = G, 
                 group = n_scnr, colour = n_scnr)) +  # group by scnr
  
  # text
  labs(title = "Stand basal area evolution",  
       # subtitle = "",  
       x = "Stand age (years)",  
       y = "Stand basal area (m²/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        # plot.subtitle = element_text(size = 15, hjust = 0.5),  
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() # lines

# watch and save graph
graph_G
ggsave(filename = '4_figures/simulation_graphs/G.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph Ho among treatments ####

graph_Ho <- 
  ggplot(df, aes(x = T, y = Ho, 
                 group = n_scnr, colour = n_scnr)) +  # group by scnr
  
  # text
  labs(title = "Stand dominant height evolution",  
       # subtitle = "",  
       x = "Stand age (years)",  
       y = "Dominant height (m)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        # plot.subtitle = element_text(size = 15, hjust = 0.5),  
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line()  # lines

# watch and save graph
graph_Ho
ggsave(filename = '4_figures/simulation_graphs/HO.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph V among treatments ####

graph_V <- 
  ggplot(df, aes(x = T, y = V_all, 
                 group = n_scnr, colour = n_scnr)) +  # group by scnr
  
  # text
  labs(title = "Stand volume evolution",  
       subtitle = "Solid line represents total stand volume (field + extracted) and dashed line represents the final stand volume",  # subtítulo
       x = "Stand age (years)",  
       y = "Volume over bark (m³/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = V), shape = 1) +  # points
  geom_line(aes(y = V), linetype = "dashed")   # lines

# watch and save graph
graph_V
ggsave(filename = '4_figures/simulation_graphs/V.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph WT among treatments ####

graph_WT <- 
  ggplot(df, aes(x = T, y = WT_all, 
                 group = n_scnr, colour = n_scnr,)) +  # group by scnr
  
  # text
  labs(title = "Stand total biomass production",  
       subtitle = "Solid line represents total stand biomass (field + extracted) and dashed line represents the final stand biomass",  # subtítulo
       x = "Stand age (years)",  
       y = "Total biomass (tn/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = WT), shape = 1) +  # points
  geom_line(aes(y = WT), linetype = "dashed")   # lines

# watch and save graph
graph_WT
ggsave(filename = '4_figures/simulation_graphs/WT.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph WR among treatments ####

graph_WR <- 
  ggplot(df, aes(x = T, y = WR_all, 
                 group = n_scnr, colour = n_scnr,)) +  # group by scnr
  
  # text
  labs(title = "Stand root biomass production",  
       subtitle = "Solid line represents total stand root biomass (field + extracted) and dashed line represents the final stand root biomass",  # subtítulo
       x = "Stand age (years)",  
       y = "Root biomass (tn/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = WR), shape = 1) +  # points
  geom_line(aes(y = WR), linetype = "dashed")   # lines

# watch and save graph
graph_WR
ggsave(filename = '4_figures/simulation_graphs/WR.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)



#### Graph WSW among treatments ####

graph_WSW <- 
  ggplot(df, aes(x = T, y = WSW_all, 
                 group = n_scnr, colour = n_scnr,)) +  # group by scnr
  
  # text
  labs(title = "Stand stem biomass production",  
       subtitle = "Solid line represents total stand stem biomass (field + extracted) and dashed line represents the final stand stem biomass",  # subtítulo
       x = "Stand age (years)",  
       y = "Stem biomass (tn/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = WSW), shape = 1) +  # points
  geom_line(aes(y = WSW), linetype = "dashed")   # lines

# watch and save graph
graph_WSW
ggsave(filename = '4_figures/simulation_graphs/WSW.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph WR among treatments ####

graph_WB <- 
  ggplot(df, aes(x = T, y = WB_all, 
                 group = n_scnr, colour = n_scnr,)) +  # group by scnr
  
  # text
  labs(title = "Stand branches biomass production",  
       subtitle = "Solid line represents total stand branches biomass (field + extracted) and dashed line represents the final stand branches biomass",  # subtítulo
       x = "Stand age (years)",  
       y = "Branches biomass (tn/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = WB), shape = 1) +  # points
  geom_line(aes(y = WB), linetype = "dashed")   # lines

# watch and save graph
graph_WB
ggsave(filename = '4_figures/simulation_graphs/WB.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)


#### Graph CARBON among treatments ####

graph_C <- 
  ggplot(df, aes(x = T, y = CARBON_all, 
                 group = n_scnr, colour = n_scnr,)) +  # group by scnr
  
  # text
  labs(title = "Stand carbon sequestration",  
       subtitle = "Solid line represents total carbon sequestered (field + extracted) and dashed line represents the final stand carbon sequestered",  # subtítulo
       x = "Stand age (years)",  
       y = "Root biomass (tn/ha)"   
  ) +
  
  # theme bw
  theme_bw() +
  
  # text position and size
  theme(plot.title = element_text(size = 25, hjust = 0.5), # title  
        plot.subtitle = element_text(size = 20, hjust = 0.5),  
        strip.text = element_text(size = 15),  # facet text
        axis.title = element_text(size = 20),  # axis
        legend.title = element_text(size = 20),  # legend title
        legend.text = element_text(size = 15)) +  # legend content
  
  # set colors and legend name manually
  scale_color_viridis('Scenario', discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE) +
  
  # plot data
  geom_point() +  # points
  geom_line() +  # lines
  
  # plot carbon data
  geom_point(aes(y = CARBON), shape = 1) +  # points
  geom_line(aes(y = CARBON), linetype = "dashed")   # lines

# watch and save graph
graph_C
ggsave(filename = '4_figures/simulation_graphs/CARBON.png', device = 'png', units = 'mm', dpi = 300, width = 300, height = 300)



#### Joint graph ####

# basic metrics

# delete previous graphs text
# graph_N <- graph_N + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0))
# graph_Ho <- graph_Ho + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0))
# graph_dg <- graph_dg + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0))
# graph_G <- graph_G + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0))

# compile graphs in one: N, Ho, dg, G
ggarrange(graph_N, graph_Ho,
          graph_dg, graph_G,
          ncol = 2, nrow = 2, 
          common.legend = TRUE, legend = "bottom",  # "bottom"
          # labels = c(' 1', ' 2', ' 3', ' 4'), 
          hjust = 0, font.label = list(size = 20))

ggsave("4_figures/simulation_graphs/joined_N_Ho_dg_G_text.png", units = 'mm', dpi = 300, width = 450, height = 300)


# biomass metrics

# delete previous graphs text
# graph_WT <- graph_WT + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
# graph_WSW <- graph_WSW + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
# graph_WB <- graph_WB + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
# graph_C <- graph_C + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
# graph_V <- graph_V + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
# graph_WR <- graph_WR + theme(plot.title = element_text(size = 0), axis.title = element_text(size = 0), plot.subtitle = element_text(size = 0))
graph_WT <- graph_WT + theme(plot.subtitle = element_text(size = 0))
graph_WSW <- graph_WSW + theme(plot.subtitle = element_text(size = 0))
graph_WB <- graph_WB + theme(plot.subtitle = element_text(size = 0))
graph_C <- graph_C + theme(plot.subtitle = element_text(size = 0))
graph_V <- graph_V + theme(plot.subtitle = element_text(size = 0))
graph_WR <- graph_WR + theme(plot.subtitle = element_text(size = 0))

# compile graphs in one: WSW, WB, WR
ggarrange(graph_WSW, graph_WB, graph_WR,
          ncol = 3, nrow = 1, 
          common.legend = TRUE, legend = "bottom",  # "bottom"
          # labels = c(' 1', ' 2', ' 3'), 
          hjust = 0, font.label = list(size = 20))

ggsave("4_figures/simulation_graphs/joined__WSW_WB_WR_text.png", units = 'mm', dpi = 300, width = 450, height = 300)

# compile graphs in one: V, WT, C
ggarrange(graph_V, graph_WT, graph_C,
          ncol = 3, nrow = 1, 
          common.legend = TRUE, legend = "bottom",  # "bottom"
          # labels = c(' 1', ' 2', ' 3'), 
          hjust = 0, font.label = list(size = 20))

ggsave("4_figures/simulation_graphs/joined_V_WT_C_text.png", units = 'mm', dpi = 300, width = 450, height = 300)
